PluginRepository - starting out
naomi/__main__.py - added arguments for working with repositories:
  --list-availabile-plugins
  --install
  --remove
  --enable
  --disable

and changed
  --list_plugins
to
  --list_installed_plugins

Right now the only option that actually works is
--list-available-plugins
and it is working against my copy of plugin.csv located at
https://github.com/aaronchantrill/naomi-plugins/blob/master/plugins.csv

naomi/application.py - added list_available_plugins method.

naomi/pluginstore.py - added a printplugin function.

naomi/strcmpci.py - function for performing case insensitive comparisons
which is an important thing for Noami to be able to do, and should be
done consistently and correctly, especially as we add languages.