SNR_VAD better hearing
I have noticed that Naomi seems to grow hard of hearing as time
goes on when using the SNR_VAD Voice Activity Detector. This
appears to be because I am updating entries by adding 1 and
dividing by 2. If the count is at 1, then the count will be
updated to 1, and that level will never be removed from the list.
over time, bumps and crashes will build up and affect the
standard deviation calculation.

This update only updates keys with a count greater than 1, so
allows older keys to time out, allowing the system to generate
a more accurate standard deviation. This should allow Naomi to
recover from hearing a loud background noise or adjust to a new
noise in the environment in only a few seconds to a minute, and
make it listen harder when things get quiet.