Fix error in unit tests
On my Debian Bookworm system, the loader returned as output from
pkgutil.get_loader() no longer has a get_filename() method. Instead
it has a _resolve_filename() method which helpfully accepts the
name of the module.

I'm not sure what changed, but this change should be compatible
with both types of systems.